+++ app/crud/realtor_reviews.py
@@ -1,7 +1,10 @@
 from fastapi import HTTPException
 
+from app.schema.types import User_Type
 from app.core.database import get_db_cursor
 from app.schema.reviews import Realtor_Reviews
+
+from app.utils.helpers import get_user_type_from_id
 
 def get_one(id: int):
     query = '''
@@ -50,6 +53,9 @@ def get_all_by_realtor_user_id(realtor_user_id: int):
         return [dict(realtor_review) for realtor_review in realtor_reviews]
 
 def create(realtor_review: Realtor_Reviews):
+    user_type = get_user_type_from_id(realtor_review.realtor_user_id)
+    if (user_type != User_Type.REALTOR):
+        raise HTTPException(status_code = 400, detail = 'Review is not for a realtor')
     query = '''
                 INSERT INTO realtor_reviews 
                     (user_id, realtor_user_id, rating, review)
@@ -75,10 +81,20 @@ def update(id: int, realtor_review: Realtor_Reviews):
                 UPDATE realtor_reviews 
                 SET 
                     rating = {}, 
-                    review = '{}'
+                    review = '{}',
+                    status = '{}'
                 WHERE id = {}
+                AND user_id = {}
+                AND realtor_user_id = {}
                 RETURNING id, updated_at
-            '''.format(realtor_review.rating, realtor_review.review, id)
+            '''.format(
+                realtor_review.rating, 
+                realtor_review.review, 
+                realtor_review.status,
+                id, 
+                realtor_review.user_id, 
+                realtor_review.realtor_user_id
+            )
     try:
         with get_db_cursor() as cursor:
             cursor.execute(query)
@@ -87,11 +103,13 @@ def update(id: int, realtor_review: Realtor_Reviews):
     except Exception as e:
         raise HTTPException(status_code = 400, detail = str(e))
 
-def delete(id: int):
+def delete(id: int, user_id: int, realtor_user_id: int):
     query = '''
                 DELETE FROM realtor_reviews 
                 WHERE id = {}
-            '''.format(id)
+                AND user_id = {}
+                AND realtor_user_id = {}
+            '''.format(id, user_id, realtor_user_id)
     try:
         with get_db_cursor() as cursor:
             cursor.execute(query)
