+++ app/crud/realtors.py
@@ -1,7 +1,10 @@
 from fastapi import HTTPException
 
 from app.schema.users import Realtors
+from app.schema.types import User_Type
 from app.core.database import get_db_cursor
+
+from app.utils.helpers import get_user_type_from_id
 
 def get_one(id: int):
     query = '''
@@ -27,6 +30,9 @@ def get_all():
         return [dict(realtor) for realtor in realtors]
 
 def create(realtor: Realtors):
+    user_type = get_user_type_from_id(realtor.realtor_user_id)
+    if (user_type != User_Type.REALTOR):
+        raise HTTPException(status_code = 400, detail = 'User is not a realtor')
     query = '''
                 INSERT INTO realtors 
                     (realtor_user_id, realtor_firm_id, first_name, last_name, email, phone, address, city, state, country, postal_code, rating, review)
@@ -59,8 +65,7 @@ def create(realtor: Realtors):
 def update(id: int, realtor: Realtors):
     query = '''
                 UPDATE realtors 
-                SET 
-                    realtor_user_id = {}, 
+                SET  
                     realtor_firm_id = {}, 
                     first_name = '{}', 
                     last_name = '{}', 
@@ -74,9 +79,9 @@ def update(id: int, realtor: Realtors):
                     rating = {}, 
                     review = '{}' 
                 WHERE id = {}
+                AND realtor_user_id = {}
                 RETURNING id, realtor_user_id, first_name, updated_at
             '''.format(
-                realtor.realtor_user_id,
                 realtor.realtor_firm_id,
                 realtor.first_name,
                 realtor.last_name,
@@ -89,7 +94,8 @@ def update(id: int, realtor: Realtors):
                 realtor.postal_code,
                 realtor.rating,
                 realtor.review,
-                id
+                id,
+                realtor.realtor_user_id
             )
     try:
         with get_db_cursor() as cursor:
@@ -99,11 +105,12 @@ def update(id: int, realtor: Realtors):
     except Exception as e:
         raise HTTPException(status_code = 400, detail = str(e))
 
-def delete(id: int):
+def delete(id: int, realtor_user_id: int):
     query = '''
                 DELETE FROM realtors 
                 WHERE id = {}
-            '''.format(id)
+                AND realtor_user_id = {}
+            '''.format(id, realtor_user_id)
     try:
         with get_db_cursor() as cursor:
             cursor.execute(query)
